convert to spec, update tests
