+++ rundeckapp/src/test/groovy/rundeck/JobsXMLCodecTests.groovy
@@ -443,101 +443,6 @@ void testDecodeBasic__no_group(){
         assertEquals "incorrect jobName", "20m", jobs[0].timeout
     }
     @Test
-    public void testDecodeBasic2(){
-
-        def xml = """<joblist>
-  <job>
-    <id>8</id>
-    <name>punch2</name>
-    <description>         dig it potato
-
-* list item
-* list item2
-
-&lt;b&gt;inline html&lt;/b&gt;
-    </description>
-    <loglevel>WARN</loglevel>
-    <context>
-      <project>zig</project>
-      <options>
-        <option name='clip' value='true' />
-      </options>
-    </context>
-    <sequence>
-        <command>
-        <exec>true</exec>
-        </command>
-        <command>
-        <exec>false</exec>
-        </command>
-        <command>
-        <exec>0</exec>
-        </command>
-        <command>
-        <script>true</script>
-            <scriptargs>true</scriptargs>
-        </command>
-        <command>
-        <script>false</script>
-            <scriptargs>false</scriptargs>
-        </command>
-        <command>
-        <script>0</script>
-            <scriptargs>0</scriptargs>
-        </command>
-        <command>
-            <scriptfile>false</scriptfile>
-            <scriptargs>false</scriptargs>
-            <errorhandler  keepgoingOnSuccess='false'>
-                <scriptfile>false</scriptfile>
-                <scriptargs>0</scriptargs>
-            </errorhandler>
-        </command>
-        <command>
-            <jobref>
-            <name>false</name>
-            <group>false</group>
-            <arg line="123"/>
-            </jobref>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>2</threadcount>
-      <keepgoing>true</keepgoing>
-    </dispatch>
-  </job>
-</joblist>
-"""
-        def jobs = JobsXMLCodec.decode(xml)
-        assertNotNull jobs
-        assertEquals  '         dig it potato\n' +
-                              '\n' +
-                              '* list item\n' +
-                              '* list item2\n' +
-                              '\n' +
-                              '<b>inline html</b>' +
-                '\n    ', jobs[0].description
-        assertEquals  8, jobs[0].workflow.commands.size()
-        assertEquals 'true', jobs[0].workflow.commands[0].adhocRemoteString
-        assertEquals 'false', jobs[0].workflow.commands[1].adhocRemoteString
-        assertEquals '0', jobs[0].workflow.commands[2].adhocRemoteString
-        assertEquals 'true', jobs[0].workflow.commands[3].adhocLocalString
-        assertEquals 'true', jobs[0].workflow.commands[3].argString
-        assertEquals 'false', jobs[0].workflow.commands[4].adhocLocalString
-        assertEquals 'false', jobs[0].workflow.commands[4].argString
-        assertEquals '0', jobs[0].workflow.commands[5].adhocLocalString
-        assertEquals '0', jobs[0].workflow.commands[5].argString
-
-        assertEquals 'false', jobs[0].workflow.commands[6].adhocFilepath
-        assertEquals 'false', jobs[0].workflow.commands[6].argString
-        assertEquals 'false', jobs[0].workflow.commands[6].errorHandler.adhocFilepath
-        assertEquals '0', jobs[0].workflow.commands[6].errorHandler.argString
-
-        assertEquals 'false', jobs[0].workflow.commands[7].jobName
-        assertEquals 'false', jobs[0].workflow.commands[7].jobGroup
-        assertEquals '123', jobs[0].workflow.commands[7].argString
-    }
-    @Test
     public void testDecodeLoglimitNoStatus(){
 
         def xml = """<joblist>
@@ -742,90 +647,6 @@ void testDecodeBasic__no_group(){
             assertEquals ("a test${i+1}".toString(),v.description)
         }
     }
-    @Test
-    public void testDecodeBasicScriptInterpreter(){
-
-        def xml = """<joblist>
-  <job>
-    <id>8</id>
-    <name>punch2</name>
-    <description>dig it potato</description>
-    <loglevel>WARN</loglevel>
-    <context>
-      <project>zig</project>
-      <options>
-        <option name='clip' value='true' />
-      </options>
-    </context>
-    <sequence>
-        <command>
-            <script>true</script>
-            <scriptargs>true</scriptargs>
-        </command>
-        <command>
-            <script>true</script>
-            <scriptinterpreter>bash -c</scriptinterpreter>
-            <scriptargs>true</scriptargs>
-        </command>
-        <command>
-            <script>false</script>
-            <scriptinterpreter argsquoted="true">bash -c</scriptinterpreter>
-            <scriptargs>false</scriptargs>
-        </command>
-        <command>
-            <script>0</script>
-            <scriptinterpreter argsquoted="false">bash -c</scriptinterpreter>
-            <scriptargs>0</scriptargs>
-        </command>
-        <command>
-            <scriptfile>false</scriptfile>
-            <scriptargs>false</scriptargs>
-            <scriptinterpreter argsquoted="false">bash -c</scriptinterpreter>
-            <errorhandler  keepgoingOnSuccess='false'>
-                <scriptfile>false</scriptfile>
-                <scriptargs>0</scriptargs>
-            </errorhandler>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>2</threadcount>
-      <keepgoing>true</keepgoing>
-    </dispatch>
-  </job>
-</joblist>
-"""
-        def jobs = JobsXMLCodec.decode(xml)
-        assertNotNull jobs
-        assertEquals  5, jobs[0].workflow.commands.size()
-
-        assertEquals 'true', jobs[0].workflow.commands[0].adhocLocalString
-        assertEquals 'true', jobs[0].workflow.commands[0].argString
-        assertEquals null, jobs[0].workflow.commands[0].scriptInterpreter
-        assertEquals false, !!jobs[0].workflow.commands[0].interpreterArgsQuoted
-
-        assertEquals 'true', jobs[0].workflow.commands[1].adhocLocalString
-        assertEquals 'true', jobs[0].workflow.commands[1].argString
-        assertEquals 'bash -c', jobs[0].workflow.commands[1].scriptInterpreter
-        assertEquals false, !!jobs[0].workflow.commands[1].interpreterArgsQuoted
-
-        assertEquals 'false', jobs[0].workflow.commands[2].adhocLocalString
-        assertEquals 'false', jobs[0].workflow.commands[2].argString
-        assertEquals 'bash -c', jobs[0].workflow.commands[2].scriptInterpreter
-        assertEquals true, !!jobs[0].workflow.commands[2].interpreterArgsQuoted
-
-        assertEquals '0', jobs[0].workflow.commands[3].adhocLocalString
-        assertEquals '0', jobs[0].workflow.commands[3].argString
-        assertEquals 'bash -c', jobs[0].workflow.commands[3].scriptInterpreter
-        assertEquals false, !!jobs[0].workflow.commands[3].interpreterArgsQuoted
-
-        assertEquals 'false', jobs[0].workflow.commands[4].adhocFilepath
-        assertEquals 'false', jobs[0].workflow.commands[4].argString
-        assertEquals 'false', jobs[0].workflow.commands[4].errorHandler.adhocFilepath
-        assertEquals 'bash -c', jobs[0].workflow.commands[4].scriptInterpreter
-        assertEquals false, !!jobs[0].workflow.commands[4].interpreterArgsQuoted
-        assertEquals '0', jobs[0].workflow.commands[4].errorHandler.argString
-
-    }
     /**
      * Empty options declaration
      */
@@ -917,105 +738,6 @@ void testDecodeBasic__no_group(){
         assertEquals "incorrect nodeThreadcount", 1, jobs[0].nodeThreadcount
     }
 
-    @Test
-    void testDecodeErrorhandler(){
-        def basic7 = """<joblist>
-  <job>
-    <id>8</id>
-    <name>punch2</name>
-    <description>dig it potato</description>
-    <loglevel>WARN</loglevel>
-    <context>
-      <project>zig</project>
-      <options>
-        <option name='clip' value='true' />
-      </options>
-    </context>
-    <sequence>
-        <command>
-            <exec>test</exec>
-            <errorhandler>
-                <exec>testerr</exec>
-            </errorhandler>
-        </command>
-        <command>
-            <script>test2</script>
-            <scriptargs>blah blah</scriptargs>
-            <errorhandler>
-                <script>test2err</script>
-                <scriptargs>blah blah err</scriptargs>
-            </errorhandler>
-        </command>
-        <command>
-            <scriptfile>test3</scriptfile>
-            <scriptargs>blah3 blah3</scriptargs>
-            <errorhandler  keepgoingOnSuccess='false'>
-                <scriptfile>test3err</scriptfile>
-                <scriptargs>blah3 blah3 err</scriptargs>
-            </errorhandler>
-        </command>
-        <command>
-            <jobref name="test" group="group"/>
-            <errorhandler keepgoingOnSuccess='true'>
-                <jobref name="testerr" group="grouperr">
-                    <arg line="line err"/>
-                </jobref>
-            </errorhandler>
-        </command>
-        <command>
-            <step-plugin type="blah">
-                <configuration>
-                    <entry key="elf" value="cheese"/>
-                </configuration>
-            </step-plugin>
-            <errorhandler keepgoingOnSuccess='true'>
-                <node-step-plugin type="blah2">
-                    <configuration>
-                        <entry key="rice" value="pilaf"/>
-                    </configuration>
-                </node-step-plugin>
-            </errorhandler>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>2</threadcount>
-      <keepgoing>true</keepgoing>
-    </dispatch>
-  </job>
-</joblist>
-"""
-        def jobs = JobsXMLCodec.decode(basic7)
-        assertNotNull jobs
-        assertEquals 1, jobs.size()
-        ScheduledExecution se=jobs[0]
-        assertEquals(5,jobs[0].workflow.commands.size())
-        jobs[0].workflow.commands.each{
-            assertNotNull(it.errorHandler)
-        }
-        assertEquals('testerr', jobs[0].workflow.commands[0].errorHandler.adhocRemoteString)
-        assertNull(jobs[0].workflow.commands[0].errorHandler.argString)
-        assertFalse(jobs[0].workflow.commands[0].errorHandler.keepgoingOnSuccess)
-
-        assertEquals('test2err', jobs[0].workflow.commands[1].errorHandler.adhocLocalString)
-        assertEquals('blah blah err',jobs[0].workflow.commands[1].errorHandler.argString)
-        assertFalse(jobs[0].workflow.commands[1].errorHandler.keepgoingOnSuccess)
-
-        assertEquals('test3err', jobs[0].workflow.commands[2].errorHandler.adhocFilepath)
-        assertEquals('blah3 blah3 err',jobs[0].workflow.commands[2].errorHandler.argString)
-        assertFalse(jobs[0].workflow.commands[2].errorHandler.keepgoingOnSuccess)
-
-        assertEquals('testerr', jobs[0].workflow.commands[3].errorHandler.jobName)
-        assertEquals('grouperr', jobs[0].workflow.commands[3].errorHandler.jobGroup)
-        assertEquals('line err',jobs[0].workflow.commands[3].errorHandler.argString)
-        assertNotNull(jobs[0].workflow.commands[3].errorHandler.keepgoingOnSuccess)
-        assertTrue(jobs[0].workflow.commands[3].errorHandler.keepgoingOnSuccess)
-
-        assertTrue( jobs[0].workflow.commands[4].errorHandler.nodeStep)
-        assertEquals('blah2', jobs[0].workflow.commands[4].errorHandler.type)
-        assertEquals([rice:'pilaf'],jobs[0].workflow.commands[4].errorHandler.configuration)
-        assertNotNull(jobs[0].workflow.commands[4].errorHandler.keepgoingOnSuccess)
-        assertTrue(jobs[0].workflow.commands[4].errorHandler.keepgoingOnSuccess)
-    }
 
     @Test
     void testDecodeExample(){
@@ -1063,74 +785,6 @@ void testDecodeBasic__no_group(){
             assertEquals "incorrect project",'demo',jobs[0].project
     }
 
-    @Test
-    void testDecodeStringsShouldNotBeBoolean() {
-        def example1 = """<joblist>
-  <job>
-    <id>1</id>
-    <name>false</name>
-    <description >false</description>
-    <loglevel>VERBOSE</loglevel>
-    <group>false</group>
-    <context>
-      <project>proj1</project>
-      <options>
-        <option name="false" enforcedvalues="false"/>
-        <option required="false" name="x" value="9000636026"/>
-      </options>
-    </context>
-    <sequence keepgoing="false">
-        <command>
-            <exec>false</exec>
-            <errorhandler keepgoingOnSuccess="false">
-                <script>false</script>
-                <scriptargs>false</scriptargs>
-            </errorhandler>
-        </command>
-    </sequence>
-    <nodefilters excludeprecedence="false">
-      <include>
-        <hostname>false</hostname>
-        <tags />
-        <os-name />
-        <os-family />
-        <os-arch />
-        <os-version />
-        <name />
-      </include>
-    </nodefilters>
-    <dispatch>
-      <threadcount>2</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-  </job>
-</joblist>
-"""
-        def jobs = JobsXMLCodec.decode(example1)
-        assertNotNull jobs
-        assertEquals "false",jobs[0].jobName
-        assertEquals  "false",jobs[0].groupPath
-        assertEquals  "false",jobs[0].description
-        assertEquals  false, jobs[0].nodeExcludePrecedence
-        assertEquals  false, jobs[0].nodeKeepgoing
-        assertEquals  null, jobs[0].nodeInclude
-        assertEquals  "hostname: false", jobs[0].filter
-        assertEquals  'proj1',jobs[0].project
-        assertEquals  1, jobs[0].workflow.commands.size()
-        assertEquals  "false", jobs[0].workflow.commands[0].adhocRemoteString
-        assertEquals  "false", jobs[0].workflow.commands[0].errorHandler.adhocLocalString
-        assertEquals  "false", jobs[0].workflow.commands[0].errorHandler.argString
-        assertEquals  false, jobs[0].workflow.commands[0].errorHandler.keepgoingOnSuccess
-        assertEquals  2, jobs[0].nodeThreadcount
-        assertEquals  false, jobs[0].workflow.keepgoing
-        assertEquals 2, jobs[0].options.size()
-        def opts=new ArrayList(jobs[0].options)
-        assertEquals 'false', opts[0].name
-        assertEquals false, opts[0].enforced
-        assertEquals 'x', opts[1].name
-        assertEquals false, opts[1].required
-        assertEquals '9000636026', opts[1].defaultValue
-    }
     @Test
     void testDecodeNodefilter() {
         /** node filter job */
@@ -1989,389 +1643,6 @@ void testDecodeBasic__no_group(){
         }catch(JobXMLException e){
 
         }
-    }
-    @Test
-    void testDecodeWorkflow() {
-        //simple workflow with script command
-        def xml6 = """<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <exec>a script</exec>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-"""
-
-            def jobs = JobsXMLCodec.decode(xml6)
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            def cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect adhocExecution: ${cmd1.adhocExecution}", cmd1.adhocExecution
-            assertEquals "incorrect adhocRemoteString", 'a script', cmd1.adhocRemoteString
-            assertNull "incorrect adhocLocalString", cmd1.adhocLocalString
-            assertNull "incorrect adhocFilepath", cmd1.adhocFilepath
-            assertNull "incorrect argString", cmd1.argString
-
-        //simple workflow with script content
-        def xml7 = """<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <script>a script 2</script>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-"""
-
-            jobs = JobsXMLCodec.decode(xml7)
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect adhocExecution: ${cmd1.adhocExecution}", cmd1.adhocExecution
-            assertEquals "incorrect adhocLocalString", 'a script 2', cmd1.adhocLocalString
-            assertNull "incorrect adhocRemoteString", cmd1.adhocRemoteString
-            assertNull "incorrect adhocFilepath", cmd1.adhocFilepath
-            assertNull "incorrect argString", cmd1.argString
-        //simple workflow with script content
-        def xml8 = """<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <scriptfile>/a/path/to/a/script</scriptfile>
-            <scriptargs>-some args -to the -script</scriptargs>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-"""
-
-            jobs = JobsXMLCodec.decode(xml8)
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect adhocExecution: ${cmd1.adhocExecution}", cmd1.adhocExecution
-            assertNull "incorrect adhocLocalString", cmd1.adhocLocalString
-            assertNull "incorrect adhocRemoteString", cmd1.adhocRemoteString
-            assertEquals "incorrect adhocFilepath", '/a/path/to/a/script', cmd1.adhocFilepath
-            assertEquals "incorrect argString", '-some args -to the -script', cmd1.argString
-        //simple workflow with jobref without jobGroup
-            jobs = JobsXMLCodec.decode("""<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <jobref name="bob" />
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-""")
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect type: ${cmd1}", (cmd1 instanceof JobExec)
-            assertNull "incorrect argString", cmd1.argString
-            assertEquals "incorrect jobName", 'bob', cmd1.jobName
-            assertNull "incorrect jobGroup", cmd1.jobGroup
-            assertEquals "incorrect nodeStep", false, !!cmd1.nodeStep
-
-        //simple workflow with jobref
-            jobs = JobsXMLCodec.decode("""<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <jobref name="bob" group="/some/path"/>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-""")
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect type: ${cmd1}", (cmd1 instanceof JobExec)
-            assertNull "incorrect adhocRemoteString", cmd1.argString
-            assertEquals "incorrect jobName", 'bob', cmd1.jobName
-            assertEquals "incorrect jobGroup", '/some/path', cmd1.jobGroup
-            assertEquals "incorrect nodeStep", false, !!cmd1.nodeStep
-
-        //simple workflow with step-first strategy
-            jobs = JobsXMLCodec.decode("""<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence strategy="step-first">
-        <command>
-            <jobref name="bob" group="/some/path"/>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-""")
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "step-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect type: ${cmd1}", (cmd1 instanceof JobExec)
-            assertNull "incorrect adhocRemoteString", cmd1.argString
-            assertEquals "incorrect jobName", 'bob', cmd1.jobName
-            assertEquals "incorrect jobGroup", '/some/path', cmd1.jobGroup
-            assertEquals "incorrect nodeStep", false, !!cmd1.nodeStep
-
-        //jobref item with args
-            jobs = JobsXMLCodec.decode("""<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <jobref name="bob" group="/some/path">
-                <arg line="-test1 1 -test2 2"/>
-            </jobref>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-""")
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect type: ${cmd1}", (cmd1 instanceof JobExec)
-            assertNotNull "incorrect adhocRemoteString", cmd1.argString
-            assertEquals "incorrect adhocRemoteString", "-test1 1 -test2 2",cmd1.argString
-            assertEquals "incorrect jobName", 'bob', cmd1.jobName
-            assertEquals "incorrect jobGroup", '/some/path', cmd1.jobGroup
-            assertEquals "incorrect nodeStep", false, !!cmd1.nodeStep
-
-        //jobref item nodeStep=true
-            jobs = JobsXMLCodec.decode("""<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <jobref name="bob" group="/some/path" nodeStep="true">
-                <arg line="-test1 1 -test2 2"/>
-            </jobref>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-""")
-            assertNotNull jobs
-            assertEquals "incorrect size", 1, jobs.size()
-            assertNotNull "incorrect workflow", jobs[0].workflow
-            assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-            assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-            assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-            cmd1 = jobs[0].workflow.commands[0]
-            assertNotNull "incorrect workflow", cmd1
-            assertTrue "incorrect type: ${cmd1}", (cmd1 instanceof JobExec)
-            assertNotNull "incorrect adhocRemoteString", cmd1.argString
-            assertEquals "incorrect adhocRemoteString", "-test1 1 -test2 2",cmd1.argString
-            assertEquals "incorrect jobName", 'bob', cmd1.jobName
-            assertEquals "incorrect jobGroup", '/some/path', cmd1.jobGroup
-            assertEquals "incorrect nodeStep", true, !!cmd1.nodeStep
-
-        //simple workflow with script content
-        jobs = JobsXMLCodec.decode("""<joblist>
-  <job>
-    <id>5</id>
-    <name>wait1</name>
-    <description></description>
-    <loglevel>INFO</loglevel>
-    <context>
-        <project>test1</project>
-    </context>
-    <sequence>
-        <command>
-            <scripturl>http://example.com/a/path/to/a/script</scripturl>
-            <scriptargs>-some args -to the -script</scriptargs>
-        </command>
-    </sequence>
-    <dispatch>
-      <threadcount>1</threadcount>
-      <keepgoing>false</keepgoing>
-    </dispatch>
-    <schedule>
-      <time hour='11' minute='21' />
-      <weekday day='*' />
-      <month month='*' />
-    </schedule>
-  </job>
-</joblist>
-""")
-        assertNotNull jobs
-        assertEquals "incorrect size", 1, jobs.size()
-        assertNotNull "incorrect workflow", jobs[0].workflow
-        assertEquals "incorrect workflow strategy", "node-first", jobs[0].workflow.strategy
-        assertNotNull "incorrect workflow strategy", jobs[0].workflow.commands
-        assertEquals "incorrect workflow strategy", 1, jobs[0].workflow.commands.size()
-        cmd1 = jobs[0].workflow.commands[0]
-        assertNotNull "incorrect workflow", cmd1
-        assertTrue "incorrect adhocExecution: ${cmd1.adhocExecution}", cmd1.adhocExecution
-        assertNull "incorrect adhocLocalString", cmd1.adhocLocalString
-        assertNull "incorrect adhocRemoteString", cmd1.adhocRemoteString
-        assertEquals "incorrect adhocFilepath", 'http://example.com/a/path/to/a/script', cmd1.adhocFilepath
-        assertEquals "incorrect argString", '-some args -to the -script', cmd1.argString
     }
     @Test
     void testDecodeWorkflowJobref(){
@@ -4480,50 +3751,6 @@ void testDecodeBasic__no_group(){
 
 
 
-    @Test
-    void testEncodeDecode(){
-        def XmlParser parser = new XmlParser()
-
-        //test multiline script content
-        def jobs3 = [
-                new ScheduledExecution(
-                        jobName:'test job 1',
-                        description:'test descrip',
-                        loglevel: 'INFO',
-                        project:'test1',
-                        //type
-                        //name
-                        //command
-
-                        workflow: new Workflow(keepgoing: true, commands: [new CommandExec(
-                            argString:'elf biscuits',
-                            adhocExecution:true,
-                            adhocLocalString:'#!/bin/bash\n\necho what is this monkey < test.out\n\necho this is a test\n\nexit 0',
-                                            
-                            )]),
-                        nodeThreadcount:1,
-                        nodeKeepgoing:true,
-                )
-        ]
-
-            def xmlstr = JobsXMLCodec.encode(jobs3)
-            assertNotNull xmlstr
-            assertTrue xmlstr instanceof String
-
-            def jobs=JobsXMLCodec.decode(xmlstr)
-            assertNotNull jobs
-            assertEquals "incorrect size",1,jobs.size()
-            def job1=jobs[0]
-        assertNotNull(job1.workflow)
-        assertNotNull(job1.workflow.commands)
-        assertEquals(1,job1.workflow.commands.size())
-        def wfi=job1.workflow.commands[0]
-            assertEquals "incorrect adhocLocalString","#!/bin/bash\n\necho what is this monkey < test.out\n\necho this is a test\n\nexit 0",wfi.adhocLocalString
-            assertEquals "incorrect argString","elf biscuits",wfi.argString
-            assertEquals "incorrect adhocRemoteString",null,wfi.adhocRemoteString
-            assertEquals "incorrect adhocFilepath",null,wfi.adhocFilepath
-            
-    }
 
     @Test
     void testEncodeLoglimit() {
